move `Textarea` into `TextBox.tsx`
