+++ packages/kiwi-react/src/bricks/TextBox.tsx
@@ -7,7 +7,6 @@ import * as Ariakit from "@ariakit/react";
 import cx from "classnames";
 import { useFieldDescribedBy, useFieldId } from "./Field.js";
 import { Icon } from "./Icon.js";
-import { Textarea } from "./Textarea.js";
 import { useMergedRefs } from "./~hooks.js";
 import { type FocusableProps, type BaseProps, forwardRef } from "./~utils.js";
 
@@ -85,6 +84,58 @@ const TextBoxInput = forwardRef<"input", TextBoxInputProps>(
 	},
 );
 DEV: TextBoxInput.displayName = "TextBox.Input";
+
+// ----------------------------------------------------------------------------
+
+interface TextareaProps extends FocusableProps<"textarea"> {}
+
+/**
+ * A styled textarea element that allows users to enter multiline text values.
+ *
+ * Example usage:
+ * ```tsx
+ * <TextBox.Textarea defaultValue="Hello" />
+ * ```
+ *
+ * Works well with the `Field` and `Label` components.
+ * ```tsx
+ * <Field>
+ *   <Label>Leave a comment, be kind</Label>
+ *   <TextBox.Textarea />
+ * </Field>
+ * ```
+ *
+ * Underneath, it's an HTML textarea, i.e. `<textarea>`, so it supports the same props, including
+ * `value`, `defaultValue`, `onChange`, and `disabled`.
+ */
+const TextBoxTextarea = forwardRef<"textarea", TextareaProps>(
+	(props, forwardedRef) => {
+		const fieldId = useFieldId();
+		const describedBy = useFieldDescribedBy(props["aria-describedby"]);
+
+		return (
+			<Ariakit.Role.textarea
+				id={fieldId}
+				{...props}
+				className={cx("🥝-text-box", props.className)}
+				aria-describedby={describedBy}
+				/**
+				 * Use an empty string as a placeholder to fix baseline alignment in Safari.
+				 * @see https://bugs.webkit.org/show_bug.cgi?id=142968
+				 */
+				placeholder={props.placeholder ?? " "}
+				render={
+					<Ariakit.Focusable
+						accessibleWhenDisabled
+						render={props.render || <textarea />}
+					/>
+				}
+				ref={forwardedRef}
+			/>
+		);
+	},
+);
+DEV: TextBoxTextarea.displayName = "TextBox.Textarea";
 
 // ----------------------------------------------------------------------------
 
@@ -198,7 +249,7 @@ const TextBoxRootContext = React.createContext<
 export {
 	TextBoxRoot as Root,
 	TextBoxInput as Input,
-	Textarea,
+	TextBoxTextarea as Textarea,
 	TextBoxIcon as Icon,
 	TextBoxText as Text,
 };
